Relocate noise.passes module files
