+++ qiskit/providers/aer/noise/passes/local_noise_pass.py
@@ -10,18 +10,15 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 """
-Noise addition passes.
+Local noise addition pass.
 """
-from typing import Optional, Union, Sequence, Callable, List, Iterable
+from typing import Optional, Union, Sequence, Callable, Iterable
 
-import numpy as np
-
-from qiskit.circuit import Instruction, QuantumCircuit
+from qiskit.circuit import Instruction
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.transpiler import TransformationPass
 from qiskit.transpiler.exceptions import TranspilerError
-from .errors import QuantumError
-from .errors.standard_errors import thermal_relaxation_error
+from ..errors import QuantumError
 
 InstructionLike = Union[Instruction, QuantumError]
 
@@ -62,12 +59,12 @@ class LocalNoisePass(TransformationPass):
         """Initialize noise pass.
 
         Args:
-            fn: noise function `fn(inst, qubits) -> InstructionLike`.
+            func: noise function `fn(inst, qubits) -> InstructionLike`.
             ops: Optional, single or list of instructions to apply the
-                                  noise function to. If None the noise function will be
-                                  applied to all instructions in the circuit.
+                noise function to. If None the noise function will be
+                applied to all instructions in the circuit.
             method: method for inserting noise. Allow methods are
-                           'append', 'prepend', 'replace'.
+                'append', 'prepend', 'replace'.
         Raises:
             TranspilerError: if an invalid option is specified.
         """
@@ -78,7 +75,7 @@ class LocalNoisePass(TransformationPass):
         if isinstance(ops, str):
             ops = [ops]
         super().__init__()
-        self._fn = fn
+        self._func = func
         self._ops = set(ops) if ops else {}
         self._method = method
 
@@ -97,7 +94,7 @@ class LocalNoisePass(TransformationPass):
                 continue
 
             qubits = [qubit_indices[q] for q in node.qargs]
-            new_op = self._fn(node.op, qubits)
+            new_op = self._func(node.op, qubits)
             if new_op is None:
                 continue
             if not isinstance(new_op, Instruction):
@@ -125,85 +122,3 @@ class LocalNoisePass(TransformationPass):
             dag.substitute_node_with_dag(node, new_dag)
 
         return dag
-
-
-class RelaxationNoisePass(LocalNoisePass):
-    """Add duration dependent thermal relaxation noise after instructions."""
-
-    def __init__(
-            self,
-            t1s: List[float],
-            t2s: List[float],
-            dt: float,
-            ops: Optional[Union[Instruction, Sequence[Instruction]]] = None,
-            excited_state_populations: Optional[List[float]] = None,
-    ):
-        """Initialize RelaxationNoisePass.
-
-        Args:
-            t1s: List of T1 times in seconds for each qubit.
-            t2s: List of T2 times in seconds for each qubit.
-            dt: ...
-            ops: Optional, the operations to add relaxation to. If None
-                 relaxation will be added to all operations.
-            excited_state_populations: Optional, list of excited state populations
-                for each qubit at thermal equilibrium. If not supplied or obtained
-                from the backend this will be set to 0 for each qubit.
-        """
-        self._t1s = np.asarray(t1s)
-        self._t2s = np.asarray(t2s)
-        if excited_state_populations is not None:
-            self._p1s = np.asarray(excited_state_populations)
-        else:
-            self._p1s = np.zeros(len(t1s))
-        self._dt = dt
-        super().__init__(self._thermal_relaxation_error, ops=ops, method="append")
-
-    def _thermal_relaxation_error(
-            self,
-            op: Instruction,
-            qubits: Sequence[int]
-    ):
-        """Return thermal relaxation error on each gate qubit"""
-        duration = op.duration
-        if duration == 0:
-            return None
-
-        # convert time unit in seconds
-        duration = duration * self._dt
-
-        t1s = self._t1s[qubits]
-        t2s = self._t2s[qubits]
-        p1s = self._p1s[qubits]
-
-        # pylint: disable=invalid-name
-        if op.num_qubits == 1:
-            t1, t2, p1 = t1s[0], t2s[0], p1s[0]
-            if t1 == np.inf and t2 == np.inf:
-                return None
-            return thermal_relaxation_error(t1, t2, duration, p1)
-
-        # General multi-qubit case
-        noise = QuantumCircuit(op.num_qubits)
-        for qubit, (t1, t2, p1) in enumerate(zip(t1s, t2s, p1s)):
-            if t1 == np.inf and t2 == np.inf:
-                # No relaxation on this qubit
-                continue
-            error = thermal_relaxation_error(t1, t2, duration, p1)
-            noise.append(error, [qubit])
-
-        return noise
-
-    def run(self, dag: DAGCircuit) -> DAGCircuit:
-        """Run the RelaxationNoisePass pass on `dag`.
-        Args:
-            dag: DAG to be changed.
-        Returns:
-            A changed DAG.
-        Raises:
-            TranspilerError: if failed to insert noises to the dag.
-        """
-        if dag.duration is None:
-            raise TranspilerError("This pass accepts only scheduled circuits")
-
-        return super().run(dag)
